+++ apis/extractor/src/handlers/pool-codes-bin/index.ts
@@ -7,7 +7,7 @@ import {
   deserializePoolsBinary,
   serializePoolsBinary,
 } from 'sushi/router'
-import { CHAIN_ID, POOLS_SERIALIZATION_INTERVAL } from '../../config.js'
+import { CHAIN_ID, POOLS_SERIALIZATION_INTERVAL } from '../../config/index.js'
 import extractor from '../../extractor.js'
 import { querySchema } from './schema.js'