+++ packages/sushi/src/config/token-maps/stables.ts
@@ -237,5 +237,5 @@ export const STABLES = {
   [ChainId.AVALANCHE_TESTNET]: [],
   [ChainId.SEPOLIA]: [],
   [ChainId.GÖRLI]: [],
-  [ChainId.CURTIS]: [],
+  [ChainId.CURTIS]: [USDT[ChainId.CURTIS], DAI[ChainId.CURTIS]],
 } as const satisfies Record<ChainId, Token[]>