+++ packages/sushi/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sushi
 
+## 3.2.4
+
+### Patch Changes
+
+- [`91fbf2861`](https://github.com/sushiswap/sushiswap/commit/91fbf2861dbebb5ca6e580124d5208a444846c4d) Thanks [@matthewlilley](https://github.com/matthewlilley)! - config
+
+## 3.2.3
+
+### Patch Changes
+
+- [#1688](https://github.com/sushiswap/sushiswap/pull/1688) [`3d9164f0d`](https://github.com/sushiswap/sushiswap/commit/3d9164f0d5fc6f832b3e2625857b64a35fb352e5) Thanks [@matthewlilley](https://github.com/matthewlilley)! - expose address module
+
 ## 3.2.2
 
 ### Patch Changes